fix: solve annotation issues in type checking
In Python <3.11, Python throws a TypeError indicating that Or operand is not supported...
