+++ src/cspark/sdk/errors.py
@@ -1,11 +1,28 @@
 from __future__ import annotations
 
 from datetime import datetime
-from typing import Any, Dict, Generic, Optional, cast
-
-from .types import TReq, TRequest, TResp, TResponse
+from typing import Any, Dict, Generic, Mapping, Optional, TypeVar, Union, cast
 
 __all__ = ['SparkError', 'SparkSdkError', 'SparkApiError']
+
+TReq = TypeVar('TReq')
+TResp = TypeVar('TResp')
+Headers = Union[Dict[str, str], Mapping[str, str]]
+
+
+class TRequest(Generic[TReq]):
+    def __init__(self, url: str, method: str, headers: Headers, body: Optional[TReq]):
+        self.url = url
+        self.method = method
+        self.headers = headers
+        self.body = body
+
+
+class TResponse(Generic[TResp]):
+    def __init__(self, headers: Headers, body: TResp, raw: str):
+        self.headers = headers
+        self.body = body
+        self.raw = raw
 
 
 class SparkError(Exception):
