+++ src/cspark/sdk/utils.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from typing import Any, cast
