+++ src/cspark/sdk/version.py
@@ -1,4 +1,4 @@
-from sys import version_info
+from sys import version_info as v
 
 __all__ = ['__version__', 'sdk_version', 'about']
 
@@ -8,7 +8,7 @@ sdk_version = __version__
 
 sdk_logger = f'CSPARK v{sdk_version}'
 
-platform_info = f'Python {version_info.major}.{version_info.minor}.{version_info.micro}'
+platform_info = f'Python {v.major}.{v.minor}.{v.micro}'
 
 about = f'Coherent Spark SDK v{sdk_version} ({platform_info})'
